requirements.txt
python-telegram-bot==20.6
requests
beautifulsoup4
python-dotenv